From brian3@brians.com:
I have reviewed the __wt_huffman_open function's error handling, and I
think the following modifications would make it safer:

* If it fails during the dynamic tree building step, it is possible that
'node' holds a subtree and 'node2' holds another subtree (which would
be later merged). So I would put
       if (node2 != NULL)
               recursive_free_node(env, node2);
to the cleanup code too.

* Sometimes there is a new node allocated and its pointer stored in
'tempnode', and it is not added to the tree yet. So the recursive free
operation would not free this one so I would also add
       if (tempnode != NULL)
               __wt_free(env, &tempnode, sizeof(WT_FREQTREE_NODE));
       (or something similar)

* To make the above safer, probably tempnode should be set to NULL in
line 316, after it is enqueued.

--HG--
branch : keith