Get rid of calls to __wt_malloc -- it's really __wt_calloc underneath, and future programmers might add memset calls to clear memory.
Make mutexes references to allocated memory instead of in-line structures.

Set a generation number in each API call and use it to free memory that is
no longer in-use by any thread in the system.

Restructure the cache code, and split it into multiple files.  There's now
an I/O server that does all of the reading for the system, and the linked
lists of pages on each hash bucket has been replaced with an array that's
grown on demand.

--HG--
branch : keith