Rename WT_ADDR_TO_ULONG -> WT_PTR_TO_ULONG, "addr" are database locations.
Get rid of calls to __wt_malloc -- it's really __wt_calloc underneath, and
future programmers might add memset calls to clear memory.

Make mutexes references to allocated memory instead of in-line structures.

Set a generation number in each API call and use it to free memory that is
no longer in-use by any thread in the system.

--HG--
branch : keith