Change the cache's hash bucket algorithm to grow without locking -- this means we can entirely separate cache draining from cache reading without locking. Restructure everything to work with 2 server threads: draining, which calls reconcile on dirty pages), and reading.   This is simpler than the previous model where the drain server and the read server had to communicate directly.
Move the lru value into the page, so we no longer need a "discard" flag
in the page, we have direct access to the lru value when the page is
"discarded".
Serialize reconciliation, Db.sync & the cache drain server can select the
same page at the same time.

--HG--
branch : keith