Change the cache's hash bucket algorithm to grow without locking -- this means we can entirely separate cache draining from cache reading without locking. Restructure everything to work with 2 server threads: draining, which calls reconcile on dirty pages), and reading.   This is simpler than the previous model where the drain server and the read server had to communicate directly.
Move the page's LRU from the cache element structure into the page: this
means we no longer need a page "discard" flag, we can just set the lru
value, and it keeps the element structure small, which is our goal for
fast search.
Add macros to set the dirty flag on a page.
Don't bother checking the WT_PAGE size, it's in-memory.

--HG--
branch : keith