Get page-modification order working: two threads deleting the same record can race, and both would schedule deletes, instead of one finding the other's change.  Add a page modification generation, which will always been seen by the other thread attempting to do a modification, causing the workQ thread to return "restart", which restarts the op at the API level.  Include stat values for the number of times a method was restarted.
The workQ doesn't need to flush the last change it makes to a page, instead
depend on the flush done before control is returned to the scheduling thread
(which is holding the hazard reference).

Re-write the page modified value comment: it doesn't require an atomic write
because only the workQ modifies it.  Leave it a separate field, for now.

--HG--
branch : keith