Delete the cache in favor of a linked, in-memory tree.  Specifically, move the WT_REF structures into the in-memory pages, and link parent's directly to their children.  This implies parents cannot be evicted while they still have subtrees in memory (including row-leaf pages with off-page duplicate subtrees).   This requires a complete re-write of almost all the "cache" code.
Rename XXX_DRAIN -> XXX_EVICT.

When we reconcile a page we can't change it's size: otherwise when we
discard the page we can't figure out if keys reference on-page data
or allocated data.

Add statistics for writing a new page (the statistics field was there,
but never updated).

--HG--
branch : keith