We can't realloc() the scratch buffers themselves, we've handed out pointers to the buffers and subsequent writing of the scratch buffer memory will write the wrong memory (specifically, the problem happens when __wt_scr_release is called, because it writes the WT_BUF->flags field which the realloc moved. Change the SESSION scratch buffer array into an array of pointers, and allocate that array separately from the WT_BUF structures themselves.
Change __wt_buf_grow to allow allocations of UINT32_MAX -- I don't expect to
see them, but there's no reason to disallow them.

Don't set the WT_BUF_INUSE flag before trying to grow the buffer, if the grow
fails, we'd leave the buffer marked in-use.

Don't return immediately if the buffer grow fails, jump to the err: label so
we get an out-of-scratch-buffers message in all failure cases.