Re-write reconciliation code to instantiate in-memory only pages during splits instead of writing new pages to disk.  The idea is a newly created internal page (created as a result of a split), or an empty page, doesn't get written to disk, instead, it's flagged to be merged into its parent when its parent is reconciled, instead of every appearing on disk.  (The reason for this is that writing newly created internal pages to disk forces the tree to increase in depth, and that's not OK: instead, the tree temporarily increases in depth until the parent is reconciled, but then usualy returns to the original depth.)
The balance of power between the on-disk and in-memory pages shifts as part
of this change: in-memory pages now own the page-type and starting record
number, that is, it should be vanishingly rare for code to reach through the
WT_PAGE->WT_PAGE_DISK reference.

Pages with subtrees can no longer be evicted, which means the eviction
thread no longer needs any special read priority or read handling: delete
WT_READ_EVICT and WT_READ_PRIORITY.

Deleted pages are now maintained in the tree until their parent is evicted;
remove the WT_PAGE_DELETED error return.

Move inlined functions that are private to bulk-load into btree/bt_bulk.c.