Split bug fixes:
If the first key of a split page is an overflow, we have to copy it, not just
the overflow structure.

Merging row-store trees after a split was wrong: for top-level pages we can
simply copy the WT_CELL for the key from the original disk image.  For merge
pages, we have a key/size pair, and we have to create the WT_CELL structure
that we copy onto the page.

The size of the allocated array of WT_REF entries for both row- and column-
store internal pages created during a split was off-by-one.

We have to copy the key out of the WT_REC_LIST memory, otherwise the next
split will simply overwrite the previous split's key value.