Initial support for the salvage method: salvage doesn't work yet, but it's at a point where I want a commit.  Specifically, this code reads the file and builds an in-memory list of possible pages, then discards any of the pages that are obviously not useful (because, in the case of a leaf page, there's another leaf page that has the same or larger key range and a larger LSN, or, in the case of an overflow page, isn't referenced by a leaf page).  This code does not merge overlapping ranges, nor does it build the in-memory page we'll use to reconcile the whole mess.
Switch to 64-bit LSNs to make LSN comparisons easier.

Add the chunk size to the WT_PAGE_DISK header -- salvage needs it to figure
out how big a piece should checksum (the alternative is to checksum bigger
and bigger pieces until we pass the maximum chunk size, or find a chunk that
checksums correctly, but that's just too painful to think about).