More stability changes:
(1) There's a problem with tracking if a page's disk blocks have been discarded
in the page using the WT_PAGE_DISK_BLOCKS flag; if a page is re-written, then
the blocks need to be discarded (clearing WT_PAGE_DISK_BLOCKS), then the page
gets new blocks (setting WT_PAGE_DISK_BLOCKS).  It's easier to free the disk
blocks in the parent's WT_REF structure as they're updated.  In other words,
if we read a page from disk, we read that page into a WT_REF structure; if
the addr/size in the WT_REF structure changes, the old values must have been
replaced, and we can discard the old blocks.

(2) We can't depend on split pages being marked "inactive".  If we flushed
memory, and caused a split, but didn't discard the pages, then there's no
reason our new split page shouldn't be marked as "in-memory" and useful.
(The alternative is to go to the read server to get to the page after it's
marked inactive, and I expect future problems down the line.)  In other
words, WT_REF_INACTIVE now means just one thing: there's no thread below
this node in the tree.  Any further checks for

Also, move the page eviction work into the wrapup function -- it's the place
that knows if it's possible to evict the page from memory for real.