Change the WT_IKEY's disk-page value-cell offset to be a key-cell offset; this might actually cost some minor performance, but it makes things a lot simpler: (1) we can now figure out if the instantiated key was an overflow key, by looking at the original cell; (2) the code to move forward to the value-cell si more consistent, we're always moving to the value-cell, instead of sometimes moving to it, and sometimes knowing where it is; (3) the WT_ROW_REF_AND_KEY_FOREACH and WT_ROW_AND_KEY_FOREACH macros go away, we never need to walk the on-disk image and in-memory page at the same time.
	Clean up the reconciliation of row-store internal pages, now
that we always instantiate internal page keys, we don't need to check
to see if they're in-memory or not.

	If we don't have any overflow items on the page, we don't need
row-store internal page disk images -- discard them after creating the
in-memory page.

	Truncate row-store internal keys to 1 byte instead of putting in
a place-holder or removing it entirely: it simplifies various things in
other parts of the code: we don't have to special case page verification,
or transforming the page from its disk image to its in-memory version.

	Re-work the page offset/pointer macros, it's a minor cleanup.