Fixes/modifications triggered by Michael's changeset 2270:2932408304ac.
Rework reconciliation to not instantiate the in-memory page until we
have the parent page and parent WT_REF reference.  I don't think there
was a bug here, but it's not OK to call the in-memory instantiation
routines without a parent reference, that's not going to work in the
long run.  This change simplifies some code, anyway -- we have to fix
up those pages, and we might as well allocate the page at the same time
we fix them up.

Change the cache-counting routines to require the page's disk image be
set -- that's what we're counting after all.  (Bug fix: when we changed
row-store internal pages to not retain the disk image, this code broke
because we started cache-counting an in-memory size that had already
been discarded.)

Get rid of __wt_off_page_size, and change __wt_off_page to use the size
found in the WT_PAGE's disk image: nothing else matters, and the page's
parent reference size (WT_PSIZE(page)) may be different from the page's
size when __wt_off_page is called as the page is being discarded.

Don't pass an address/size pair to __wt_page_free -- they can only be
used for debugging information, and there's a strong temptation to use
them for other things.