A rework of Michael's fix in changeset 2269:c4e8dc1a7637.
We shouldn't depend on __rec_cell_build_key() writing data in the correct
order such that we can pass it the initial key data in the same buffer in
which it constructs the final key cell.   Use a temporary buffer instead,
and revert __wt_buf_set to using memcpy instead of memmove.