Don't limit the table name argument to 100 characters minus "table:", allocate a buffer.
Reset "ret" to 0 if at the end of the cursor walk we get WT_NOTFOUND,
else we'd exit with a failure status.

Check all calls to cursor->get_key/get_value for failure.

Check all calls to write to stdout for failure.

Fix cut-and-paste errors:
Fix the salvage utility to call the salvage method, not the verify method.
Fix the verify utility to have "verify" in its error message.