SERVER-3280: Fixing the RPM init.d file to work on Centos/Redhat
The killproc macro on these distrobutions does not support the given
parameters. Instead it supports a different set of parameters. Also
the lock file specified does not include a pid so it will not work
when killing the process anyway. Have not tested the original on
another OS so cannot verify it works on any other distro.