Add support for extending column-store files -- this isn't right yet, but it mostly works.
New configuration options: btree_column_{leaf,internal}_extend that
specify how much to extend leaf and internal pages at a time.

Deleted column-store values used to be special, on-disk page values.
That still is correct, but in addition, a 0 value for the on-page
reference means a deleted record, so that we can allocate an array of
0's and have it work.

Make the serial functions inline functions instead of #defines.  (The
#defines were getting complicated, and hard to debug, and this means we
can set/test the return value in all the normal ways.)
	Add magic to the serial functions so the workQ caller no longer
has to figure out if the workQ used the memory it was provided, instead
the memory is automatically free'd when we return to the calling thread,
before the inline function returns.

Rename __WT_PAGE_SPLIT to be __WT_PAGE_MERGE (internal column-store
pages created during file extension should be merged so the tree does
not deepen permanently, but those pages aren't the results of splits.)

Display the page's parent during debugging (it's easier to track pages).

--HG--
extra : rebase_source : 7f2589dc6b0c36a5c92a99de853adc10a2dcd96c