Add memory-footprint field to the in-memory page structure, it's used to track the memory in use by that page.  (Delete the page's cache-counted flag, if the page has memory associated with it, it's treated as cache counted.)
The workQ now has its own pages/bytes allocated fields so we can
track memory use when pages are created in memory.

Couple workq setting/incrementing the memory allocated to an in-memory
page to the serialization functions: if you pass in a chunk of malloc'd
memory, you have to pass in its logical size as well, and that size is
added to the page's memory footprint if the chunk is used.

Now that we allocate new pages in memory, we have to check to see if
the eviction server needs to run regardless of whether or not we've
read from the backing file.

Split "evict" verbose keyword into "evictserver" and "reconcile".

Include btree.i and cell.i in wt_internal.in.

--HG--
rename : src/include/api.h => src/include/cursor.h
extra : rebase_source : 0f24170ef1eee54dd7a34be45d65e96b69461af5