The eviction thread has its own structure that's used for reconciliation, it's not OK for other threads of control to use it.  First fix: verification does not need to discard the root page explicitly (close will do that for us), nor does it need to call reconciliation explicitly, just let the eviction thread do the work.
The only thing this changes is we're no longer discarding verified pages as
soon as we're done with them, we're letting the eviction server time them
out -- it's easy to rationalize this as correct behavior, if the application's
workflow is to verify then open for work, this will leave the pages in memory,
which is faster.  Yes, that assumes the API code closes the btree handle as
soon as the last reference goes away, but that's its problem, not the btree
engine's.