Rework compression error handling:
Running out of memory is expected because WiredTiger passes the minimum
amount of memory it cares about (if compression isn't going to help,
WiredTiger wants to fail fast and use the uncompressed version of the
page).  Add a flag to the compress interface which returns the fact of
compression's failure as a side-effect, and change the API so the return
from the compression function in that case is 0, to minimize the
caller's error handling.

Remove the requirement that compression/decompression return how much
memory is needed to compress/decompress a chunk.  Assuming the engine
stores the original byte string length in its header information, the
decompression code can certainly return the right value, but there's no
way an engine can bound the memory required for compression without
either picking an unreasonably big value or doing the actual
compression.  Regardless, WiredTiger doesn't want either of these
values, so simplify the interface.