Make the cursor->search_near method's "exact" argument work.
Add cursor->search function for Btree tables, cursor->search_near has
to do quite a bit more work if there's no exact match, we don't want
to slow down normal lookup.

Replace WT_CURSOR_BTREE->match with WT_CURSOR_BTREE->compare which
matches the semantics of the exact parameter to cursor->search_near,
that is, -1 if it's a smaller key in the tree, 0 for an exact match,
and 1 if it's a larger key in the tree.

The cursor->search_near function has to return a key if the match is
not exact: change __wt_return_value() to optionally return a key, and
rename it to __wt_kv_return().

Split read statistics into read and read-near.

Move set of WT_CURSOR_BTREE->flags(WT_CBT_SEARCH_SET) down into the
search routines.

Change __cursor_deleted() to be a boolean, not returning an error value,
we don't necessarily return error on deleted records.

The search routines don't return WT_RESTART, that only happens if the
tree is modified: remove restart loops from search calls, add restart
loops to modify calls.