Add a checksum to the file description block.
Don't salvage files where the file description block is corrupted, that
is, don't salvage files if the checksum, magic, major and minor numbers
don't match.  Add a force flag to the session->salvage method (available
as -F from the wt salvage command), that overrides this.

Change salvage to always overwrite the file description block, in case
it was corrupted.