Add the "multithread" connection configuration option, which configures multithreaded behavior.   Single-threaded behavior is the default, with all page update operations done by the running thread, only reads and evictions are routed through the workQ thread.  The reason is performance: the out-of-the-box performance needs to be as fast as the other OpenSource database engines, and they're single-threaded for the most part.
Change the serialization functions to return void, that is, all error
status is returned through session->wq_ret, nothing is returned from the
serialization function itself.  There are two associated changes: first,
the read/evict serialization functions weren't calling the wrapup
function on error, which meant a thread calling them might hang; second,
some of the other serialization functions weren't setting
session->wq_ret on error, which would lead to all sorts of bad behavior,
because the return from the serialization functions was ignored by the
workQ thread.  By making the serialization function return void, I'm
hoping that we can avoid this in the future, there's only one error you
can return.

Delete the "spin" flag to the serialization functions: if the op is a
WT_WORKQ_FUNC op, it means it's a non-blocking operation and we spin,
(or, in the case of single-threaded operation, we can do the work in the
thread of control that's running).

Rename __wt_connection_config -> __wt_connection_init, it's a function
to initialize the structure, not configure anything.

Change __wt_connection_destroy to return a void, it has no error returns,
this simplifies the code, as well.

--HG--
extra : rebase_source : e7398c10d250ec0321fe8248352ec8252d5b0dbc