Replace internal_node_min/leaf_node_min (the minimum page size) configuration with internal_overflow_size/leaf_overflow_size (explicit configuration of the size of an overflow item.
Fix a bug where we never actually paid any attention to the internal
overflow item size, even though we set it.

Delete the split_min debugging configuration string: the problem is that
small pages now have to be set in concert with the overflow size, if
there is an overflow size that's relatively large, and we create a split
page smaller than an item which is itself smaller than the overflow
size, the item will never fit on the page and we loop forever.  I
haven't chased any bugs in reconciliation's page size in quite awhile,
hopefully split_min is no longer interesting.

Write a basic performance tuning page for the documentation.

--HG--
extra : rebase_source : 0825ab4ff58b2dcb18675290f299fad0f5eaaae7