Make column-store search function naming match row-store (rename __col_insert_search() to __col_insert_search_match(), and rename __col_insert_search_stack() to __col_insert_search()).  Minor cleanups in those functions to always test if the WT_INSERT_HEAD argument is NULL.
The column-store insert code didn't catch the fact that there was a
race: if we find the record that's being inserted, then we know another
thread of control must have inserted it, and so we return WT_RESTART.
Found by inspection, we've never run any of this code multi-threaded.

Bug fix in the row-store search code, don't test if the address of
inshead->tail[0] is NULL, test inshead->tail[0] itself.