I was using the cbt->vslot value to decide if I needed to switch to a different insert list in variable-length column sort, and that's not good enough, just because vslot hasn't changed doesn't mean we're on the same insert list.  (For example, a cursor at some position which then does next, then prev, then next again, can use the return information for both "next" calls, but must switch insert lists on every call.)
Fundamentally, to make the current approach work, we need a variable
that flags the insert list we're on and a variable that flags the slot
into which we've configured the return information, and we have to
set/clear it appropriately on every call.

Rather than try and make this work, I switched both fixed- and
variable-length column-store next calls to match the prev calls, and
they now search the full insert list on every call,  using the skiplist.
This is going to be slower if, for example, there was an RLE entry with
a lot of elements and most of them were updated, but I don't want to put
in a bunch of complexity until we see that workload.