Fill in the rest of the statistics cursor methods, adding prev, last, set-key, search and search-near.
Change the statistics cursor "key" to be the description field of the
statistic (the previous key was the internal name, and while the
description field isn't a great choice because it might easily change,
the internal name is completely useless).

Change the order of the raw statistics return from v,pval to pval,v,
so it matches the printable return.

Add the -p option to the "wt stat" command, it prints out all stats
with the specified prefix.