Another run at memory-barriers
There were a few places where we didn't have a write-barrier between
structure content updates and publishing references to the structures.

A bug fix in __wt_session_serialize_func(): add a read-barrier before
we load  the wq_ret field after the workQ publishes the wq_state field.

A bug fix in __wt_session_serialize_wrapup(): for the search code
algorithm (read the write-generation, read-barrier, read the page) to
work, the workQ has to flush all of the page changes to memory before
it updates the write-generation, otherwise it would be possible for the
new write generation to hit memory first, and the reader to read the
updated write-generation value but not read the updated page structures.

--HG--
extra : rebase_source : 418e5f6648c9c40eebeb0a3c1339a22c05ea4be0