SERVER-3771 -- part 3 -- make shell command line behave more like bash
Remove the all-on-one-line command line editing in the shell and replace it
with something more in the style of other command lines such as bash.  Allow
prompts to exceed the width of the terminal screen and allow newlines in
them.  Strip all other control characters from the prompt text before
displaying it so that embedded backspaces, tabs and other controls don't
display badly or require extra code to emulate.  This prevents users from
using escape sequences in their prompts to perform highlighting, for now
at least.  Allow command line text to exceed the available width following
the prompt by wrapping onto other lines the way they do in bash.  Make the
newline character display as a newline instead of as a linefeed.  Increase
the amount of common code between Linux and Windows versions and make them
behave the same way.  Bracket highlighting is now working in Windows as
well as in non-Windows versions.  Don't refresh the Windows command line
when the shift key is pressed.  In Windows, when ctrl-L is pressed, try
to preserve scrollback memory if possible.  This needs additional work in
the case where the screen buffer length is less than twice the window
height.