SERVER-4008 -- don't leak memory in the shell
Track the strdup-ed (i.e. malloc-ed) strings returned by linenoise, and
free the memory when we are done with it.  Free command line history
memory explicitly when we save the history to a file and stop using it.
Non-Windows versions were freeing this memory using an atexit handler,
but this handler isn't set up for Windows.