SERVER-3771 -- parts 1 and 2 (reissue) -- do not let 'prompt' become undefined
This is a reissue of my earlier fix.  I'm trying to break them down into
bite-sized parts to make them easier to pull.

Test the return value from a call to a function prompt (included regex)
and only use it if it is a string.  If we have no string prompt, display
the default prompt = sayReplSetMemberState()+"> "; .