SERVER-3470 (reissue) -- store multi-line commands in history
This is a reissue of my earlier fix. I'm trying to break them down into
bite-sized parts to make them easier to pull.

Convert newlines to spaces and store the entire multi-line command instead
of just the line that began the command.  Fix initial cap in error message.
Fix incorrect test for failure of call to string.find().