__wt_config_concat should just concatenate the config strings.
It should not call __wt_config_get, which resolves the final value of each key:
just write the key/value pairs into the buffer and let parser that deal with
the result resolve the values.  The previous example of concatenating
"A=,A=,A=" with "A=value" should result in "A=,A=,A=,A=value".