Change the row-store leaf-page binary search code to use keys from the disk image if they're available.
Change the forward-scan cursor code to build new keys from previously returned
keys instead of rolling forward from an instantiated key.

Instead of instantiating the keys at binary set points on row-store leaf pages
when we read the page into memory, do it on demand when required by the binary
search routine.  If forced to instantiate a key, we don't just copy into a
local buffer, we instantiate the key on the page so future searches will find
it.  This has two side effects: First, if you search a Huffman-encoded page,
you end up instantiating not just the binary search points on the page, but
the keys in the final bucket you search.   This would lose if you never searched
for the same key/value pair again, but would win if you searched for the same
key/value pair again.  Second, if you only search for a single key/value pair
on the page, and never search again, we don't instantiate as many keys, as we
only instantiate the one search path, not every binary search point.

If we're doing reverse cursor scans, we have to instantiate some set of keys on
the page, otherwise we have N-squared behavior as we repeatedly roll-forward
from the first key on the page to every key/value pair on the page.  Instead,
when we move to a new page in the reverse cursor code, instantiate the keys at
the binary set points.

If Huffman encoding is configured for keys, instantiate the keys on the leaf
page as we read it in.   Here's the problem: on a Huffman encoded page, we don't
have any fast-path solutions, we have to decode the key and then figure out the
prefix encoding.   To avoid the N-squared behavior, we instantiate some set of
keys up-front, we'll need them regardless of whether we're going to do random
searches or cursor scans on the page.

--HG--
extra : rebase_source : 392dd046bd249d29cd3fce0348056959aa90e9da