copy the whole chunk for key data
doMergeChildren is changed accordingly, we merge two nodes in two separate
steps, the first step is to copy all the key data, and the second step is to
copy the key node.

In this way, we're guaranteed to have enough space on left node, otherwise,
it's possible plain pushBack() could have more data than it looks, due to the
introduction of de-dup keys.