Clean up handling of the page's reconciliation flags (WT_PAGE_REC_EMPTY, _REPLACE and _SPLIT): they no longer mean anything other than "this is the result of the last reconciliation, that is, they simply describe the WT_PAGE_MODIFY structures that are associated with the page.
This means we don't make judgements as to whether or not eviction or
reconciliation should be performed in the eviction server code, it's job
is to get a page and call the correct routine, and the underlying
routine decides what work is done (or not done) on the page.

This also means we no longer clear the WT_PAGE_REC_EMPTY flag when we
modify a page (although that isn't a bug, but it doesn't buy us enough
for the additional complication).  Which means we can't evict a page
just because the empty flag is set on its children, it might have been
dirtied since the empty flag was set.   Clearing the empty flag in the
"set the dirty flag on this page" routine handled that case, but it
doesn't not handle the case of a split page that was subsequently
dirtied, and I'm not interested in trying to fix both of them as a
result of modifying any particular page.

Break out the common code used by both row- and column-store eviction
to walk the tree and make sure the entire subtree can be evicted.

There are also a couple of minor fixes to the merge, basically,
incorrect code that didn't get spotted during the merge but the types
matched and so it compiled anyway.

--HG--
extra : rebase_source : 34b391806e2e0414a59aac010a5593b6d43b59a6