Force eviction of pages that become more than 20 times larger in memory than they are expected to be on disk.
The intent here is to smooth out performance with large caches: without this change, there is effectively no bound on in-memory page size with a large cache, and it is likely for large pages to be hotter (with uniform distribution of keys), and therefore less likely to be evicted, hence growing bigger.  Reconciling a 1GB page slows things down when the maximum page size is 16KB...

A risk is that if compression is very effective, or with off-page keys, the memory footprint can be significantly larger than the on-disk footprint even for clean pages.  This may need to become another eviction tuning parameter.