Rework verify to work in the new, block-manager world.
Rewrite the move of the root addresses into the schema file.  More
changes to the WT_BTREE handle code.  The root address is no longer
stored in any form, and is explicitly set by the eviction thread when
it pushes out the root page.  Fix bugs in our tracking of the root
page's blocks when reconciling the root page during a split.

Move the free-list maintainance from the engine into the block manager
code.

Fix bugs in the block-manager salvage support, we weren't handling
checksums correctly.   Fix bugs in salvage support, we weren't sorting
the overflow addresses correctly.

Switch the "allocate" verbose keyword to "block".