Switch from uint32_t/uint32_t addr/size pairs to off_t/uint32_t file-offset/size pairs.
Divide the size by the allocation unit as we do the offset (previously
we only reduced the addr of the addr/size pair, but the size is
divisible by the allocation unit, too.

Change verbose messages to output a session and file name if one's
available, don't output the verbose flag's name any more, the session
name usually gives the same information.