SERVER-4584 add to illegal dbname characters list
These newly added characters are illegal for Windows filenames (although it is possible to create files
with such names on NTFS, the API to access these files will prohibit them unless you do something special).
We are prohibiting these same characters on Linux because they're a bad idea to use.
Upgrading a database with a name containing illegal characters will cause the new mongod to fail to start.