avoid unit test hangs on exit
If mongo.exe exits on Windows while a terminal stream is in use, the
internal critical section of the ostream gets destructed out from
underneath it.  This causes the output thread to hang waiting for a cs
that will never become unlocked.  Switching to printf() avoids this issue.
This problem readily occurs when mongo.exe spawns other programs and pipes
their output through itself to the terminal. This is a common thing to do
in unit tests.