Always instantiate IKEY->cell_offset key-cell references in row-store internal pages (we were previously leaving them set to 0 unless it was an overflow key, and that's mildly sleazy).   Change reconciliation of row-store internal pages to set a local variable based on the unpacked cell instead of depending on that trick.
Change track initialization to return an error if it finds an incorrect
entry (there was a bug, found by inspection, where a cut-and-paste left
an assert that would never trigger).

Minor cleanup: we don't need to call __wt_get_addr() when cleaning up
after split pages, we know exactly what we have.

Update a couple of comments.