I just made a change to the eviction server code to not put both a parent and a to-be-merged child (that is, a child which will be merged into its parent during reconciliation), on the eviction list.  The problem that was supposed to fix was a core dump when both were put on the list, then the parent was evicted (merging the child), and then we attempted to evict the child (which had been free'd already).
The specific change was to not put pages marked for merge on the
eviction list at all.

However, that implies that a page marked for merge will never be
evicted, and if it's dirtied after being marked for merge, the parent
can no longer be evicted (because it references a dirty page), and the
page can no longer be evicted (because it's marked for merge), in other
words, marking a page for merge and then dirtying it  would pin the
parent/child-page pair in memory.

To fix this, I'm changing eviction to write dirty child pages found
during the walk to acquire exclusive locks on all of the subtree pages.