The change to allow applications to turn off checksums broke salvage: I changed the block-manager read code to ignore checksums of 0.  So, if salvage reads a 512B chunk in the file with a disk-size of some big number and a checksum of 0, we read the rest of the chunk, and because the checksum is 0, we treat it as a valid chunk and return it to the Btree salvage code.  The Btree salvage code figures out it's corrupted, but it's too late, we've already decided to skip the rest of the chunk down in the block-manager code, and so we miss valid blocks that were contained in that chunk.
For now, fix this by using a "random" bit pattern instead of 0 to mean
"no checksum".  There are two problems with this: First, we still have
a 1-in-2^32 chance of the problem happening anyway, all we're saying is
that a checksum of 0 is a much more likely thing to find than a checksum
of our bit pattern.  Second, there is now a 1-in-2^32 chance of a valid
checksum being ignored, that is, a page that happens to have a real
checksum that matches our magic pattern will now have the checksum
ignored, and if that page gets corrupted, we wouldn't catch it.
(Obviously, the 1-in-2^32 chance is based on uniform distribution of
checksums, and I have absolutely no idea how valid that is.)