More eviction locking cleanups and fixes.
To avoid a freed page pointer ending up on the request list, check the page state while holding the LRU lock before putting it one the request list.

In __rec_excl_page(), we can write dirty pages after we acquire the exclusive lock  -- if we can't get an exclusive lock, it means somebody else is in that part of the tree, so there's no urgency in doing the write, and if we can get the exclusive write, then we know the page will be clean when we get to the end.

Instead of copying all hazard references in the system in order to check for a conflict, just do a linear search for the reference we are interested in.

Take more care with forced eviction: keep the page in either WT_REF_EVICTING or WT_REF_LOCKED states so that application threads drain and eviction is not starved.