There's a race between the page eviction thread updating the parent's page reference (unblocking the waiting readers/writers), and the page eviction thread discarding the page, including the tree's append list.
We can fix the race by moving the tree's append list discard code into
the eviction code so the tree's append list is discarded before we permit
readers/writers to proceed (bad because we don't want page discard code
in the eviction code and because readers/writers will wait longer than
they'd otherwise wait), or by moving the tree append list into the
page's modification information (bad because we're wasting 8 bytes in
every column-store page that is being modified and isn't the last page
in the tree).

This commit does the latter, mostly because it gets rid of the page's
"last page" flag and seems cleaner, plus, I suspect it's less likely
to be involved in future bugs.